Make organisation upgradeable by coding to an interface
The Parent contract is now coded to an Organisation interface
rather than its implementation. This enables upgrading of the
Organisation contract as long as the ABI doesn't change. This also
implies that Parent can no longer create Organisations. In this
implementation, Organisation needs to be instantiated by the user
and passed on to the Parent contract for registering or upgrading.